requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
apscheduler==3.10.4
python-whois==0.9.4
pytrends==4.9.2
nltk==3.8.1
jinja2==3.1.4
sendgrid==6.11.0
tldextract==5.1.2
pydantic-settings==2.2.1
aiofiles==23.2.1
